Icons and confirmation link inserted (delete function)

+++ webapp/public/javascripts/shoppinglist.mod.households.js
@@ -52,11 +52,12 @@ var Shoppinglist = Shoppinglist || {};
         for (i = 0, len = data.households.length; i < len; i++) {
             current_item = data.households[i];
             tmpHtml.add('<li>');
+            tmpHtml.add('<img src="images/icon_house.png" height="20" width="20" border="0">');
             tmpHtml.add('<a href="controller_proxy.php?controller=fetchshoppinglists&amp;hid=' + current_item.householdId + '" hid="' + current_item.householdId + '">' + current_item.name + '</a>');
 
             if (parseInt(current_item.isOwner, 10) === 1) {
                 tmpHtml.add('<span class="actions">');
-                tmpHtml.add('<a class="delete" href="controller_proxy.php?controller=deletehousehold&amp;hid=' + current_item.householdId + '" hid="' + current_item.householdId + '">[delete]</a>');    
+                tmpHtml.add('<a class="delete" onclick="javascript:return confirm(\'Are you sure?\')" href="controller_proxy.php?controller=deletehousehold&amp;hid=' + current_item.householdId + '" hid="' + current_item.householdId + '">[delete]</a>');    
                 tmpHtml.add('<a class="invite" href="?hid=' + current_item.householdId + '" hid="' + current_item.householdId + '">[invite]</a>');
                 tmpHtml.add('</span>');
             }
